% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agitated.R, R/package.R
\docType{package}
\name{agitated}
\alias{agitated}
\title{An alternative to the UpSetR package for upset plots.}
\usage{
agitated(
  x,
  nsets = 20,
  exclusive = TRUE,
  intersection_order = c("frequency", "degree"),
  sort_sets = TRUE,
  title = NULL,
  subtitle = NULL
)
}
\arguments{
\item{x}{A list or presence/absence matrix.}

\item{nsets}{Maximum number of sets to be shown.}

\item{exclusive}{Should the intersections shown be exclusive? If yes, each 
entry is shown only once in the top bar plot.}

\item{intersection_order}{Sort by "frequency" or "degree"?}

\item{sort_sets}{Controls whether the input sets are re-ordered based on 
descending size.}

\item{title, subtitle}{Plot title and subtitle.}
}
\value{
Whatever cowplot::plot_grid returns
}
\description{
A lightweight (and reasonably efficient) alternative to the UpSetR 
package. These plots allow visualisation of an arbitrary number of set 
interactions (see http://doi.org/gbjxr9).
}
\section{agitated functions}{

agitated: produce an upset plot
}

\examples{

data <- agitated:::example_data()
agitated(data)

agitated(data, nsets = 10)

agitated(data, exclusive = FALSE)

agitated(data, intersection_order = "degree")

agitated(data, sort_sets = FALSE)

agitated(data, title = "title", subtitle = "subtitle")
}
