% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agitated.R
\name{agitated}
\alias{agitated}
\title{An alternative to upset plots}
\usage{
agitated(x, nsets = 20)
}
\arguments{
\item{x}{A list}

\item{nsets}{Maximum number of sets to be shown.}
}
\value{
Whatever cowplot::plot_grid returns
}
\description{
An alternative to upset plots
}
