% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agitated.R, R/agitated_package.R
\docType{package}
\name{agitated}
\alias{agitated}
\alias{agitated-package}
\title{An alternative to upset plots}
\usage{
agitated(x, nsets = 20, sort_by = c("size", "degree"),
  exclusive = TRUE)
}
\arguments{
\item{x}{A list}

\item{nsets}{Maximum number of sets to be shown.}

\item{sort_by}{Sort by "degree" or "size"?}

\item{exclusive}{Should the intersections shown be exclusive? If yes, each 
entry is shown only once in the top bar plot.}
}
\value{
Whatever cowplot::plot_grid returns
}
\description{
A lightweight (and reasonably efficient) alternative to the UpSetR 
package. These plots allow visualisation of an arbitrary number of set 
interactions (see http://doi.org/gbjxr9).
}
\section{agitated functions}{

agitated: produce an upset plot
}

